% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{add_functions_by_alias}
\alias{add_functions_by_alias}
\title{add_functions_by_alias}
\usage{
add_functions_by_alias(module, as = NULL)
}
\arguments{
\item{module}{Module name}

\item{as}{Alias for module name}
}
\description{
Add functions when a module name is aliased
}
\details{
Translate a function names
}
