% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{cache_eng_python}
\alias{cache_eng_python}
\title{A reticulate cache engine for Knitr}
\usage{
cache_eng_python(options)
}
\arguments{
\item{options}{List of chunk options provided by \code{knitr} during chunk execution.
Contains the caching path.}
}
\description{
This provides a \code{reticulate} cache engine for \code{knitr}. The cache engine
allows \code{knitr} to save and load Python sessions between cached chunks. The
cache engine depends on the \code{dill} Python module. Therefore, you must have
\code{dill} installed in your Python environment.
}
\details{
The engine can be activated by setting (for example)\preformatted{knitr::cache_engines$set(python = reticulate::cache_eng_python)
}

Typically, this will be set within a document's setup chunk, or by the
environment requesting that Python chunks be processed by this engine.
}
