% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{capture_stdout}
\alias{capture_stdout}
\title{Capture and return Python stdout}
\usage{
capture_stdout(expr)
}
\arguments{
\item{expr}{Expression to capture stdout for}
}
\value{
Character vector with output
}
\description{
Capture and return Python stdout
}

