% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda-params}
\alias{conda-params}
\title{conda-params}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}

\item{forge}{Boolean; include the \href{https://conda-forge.org/}{conda-forge}
repository?}

\item{channel}{An optional character vector of conda channels to include.
When specified, the \code{forge} argument is ignored. If you need to
specify multiple channels, including the conda Forge, you can use
\verb{c("conda-forge", <other channels>)}.}

\item{packages}{A character vector, indicating package names which should be
installed or removed. Use \verb{python=<version>} to request the installation
of a specific version of Python.}

\item{...}{Optional arguments, reserved for future expansion.}
}
\description{
conda-params
}
\keyword{internal}
