% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_binary}
\alias{conda_binary}
\alias{conda_exe}
\title{Find a Conda Executable}
\usage{
conda_binary(conda = "auto")

conda_exe(conda = "auto")
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}
}
\description{
Locate a conda installation on the machine, and find the associated path to
the \code{conda} executable within that installation.
}
\section{Finding Conda}{


Most of \code{reticulate}'s conda APIs accept a \code{conda} parameter, used to control
the \code{conda} binary used in their operation. When \code{conda = "auto"},
\code{reticulate} will attempt to automatically find a conda installation.
The following locations are searched, in order:
\enumerate{
\item The location specified by the \code{reticulate.conda_binary} \R option,
\item The location specified by the \code{RETICULATE_CONDA} environment variable,
\item The \code{\link[=miniconda_path]{miniconda_path()}} location (if it exists),
\item The program \code{PATH},
\item A set of pre-defined locations where conda is typically installed.
}

To force \code{reticulate} to use a particular \code{conda} binary, we recommend
setting:\preformatted{options(reticulate.conda_binary = "/path/to/conda")
}

This can be useful if your conda installation lives in a location that
\code{reticulate} is unable to automatically discover.
}

\seealso{
Other conda-tools: 
\code{\link{conda_create}()},
\code{\link{conda_export}()},
\code{\link{conda_install}()},
\code{\link{conda_list}()},
\code{\link{conda_python}()},
\code{\link{conda_remove}()},
\code{\link{conda_version}()}
}
\concept{conda-tools}
