% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_clone}
\alias{conda_clone}
\title{Clone a Conda Environment}
\usage{
conda_clone(envname, ..., clone = "base", conda = "auto")
}
\arguments{
\item{clone}{The environment to be cloned.}
}
\description{
Clone a Conda Environment
}
