% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_export}
\alias{conda_export}
\title{Export a Conda Environment}
\usage{
conda_export(
  envname,
  file = if (json) "environment.json" else "environment.yml",
  json = FALSE,
  ...,
  conda = "auto"
)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{file}{The path where the conda environment definition will be written.}

\item{json}{Boolean; should the environment definition be written as JSON?
By default, conda exports environmentas as YAML.}

\item{...}{Optional arguments, currently ignored.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}
}
\value{
The path to the exported environment definition, invisibly.
}
\description{
Export a conda environment definition, either as YAML (the default)
or JSON. The resulting environment file can be used by \code{\link[=conda_create]{conda_create()}}
to create a "clone" of the exported conda environment.
}
\seealso{
Other conda-tools: 
\code{\link{conda_binary}()},
\code{\link{conda_create}()},
\code{\link{conda_install}()},
\code{\link{conda_list}()},
\code{\link{conda_python}()},
\code{\link{conda_remove}()},
\code{\link{conda_version}()}
}
\concept{conda-tools}
