% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_install}
\alias{conda_install}
\title{Install a Package in a Conda Environment}
\usage{
conda_install(
  envname = NULL,
  packages,
  forge = TRUE,
  channel = character(),
  pip = FALSE,
  pip_options = character(),
  pip_ignore_installed = FALSE,
  conda = "auto",
  python_version = NULL,
  ...
)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{packages}{A character vector, indicating package names which should be
installed or removed. Use \verb{python=<version>} to request the installation
of a specific version of Python.}

\item{forge}{Boolean; include the \href{https://conda-forge.org/}{conda-forge}
repository?}

\item{channel}{An optional character vector of conda channels to include.
When specified, the \code{forge} argument is ignored. If you need to
specify multiple channels, including the conda Forge, you can use
\verb{c("conda-forge", <other channels>)}.}

\item{pip}{Boolean; use \code{pip} for package installation? By default, packages
are installed from the active conda channels.}

\item{pip_options}{An optional character vector of additional command line
arguments to be passed to \code{pip}. Only relevant when \code{pip = TRUE}.}

\item{pip_ignore_installed}{Ignore already-installed versions when using pip?
(defaults to \code{FALSE}). Set this to \code{TRUE} so that specific package versions
can be installed even if they are downgrades. The \code{FALSE} option is useful
for situations where you don't want a pip install to attempt an overwrite of
a conda binary package (e.g. SciPy on Windows which is very difficult to
install via pip due to compilation requirements).}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}

\item{python_version}{The version of Python to be installed. Set this if
you'd like to change the version of Python associated with a particular
conda environment.}

\item{...}{Optional arguments, reserved for future expansion.}
}
\description{
Install a package into a conda environment. Packages can be installed either
from the conda repositories, or via \code{pip}.
}
\seealso{
Other conda-tools: 
\code{\link{conda_binary}()},
\code{\link{conda_create}()},
\code{\link{conda_export}()},
\code{\link{conda_list}()},
\code{\link{conda_python}()},
\code{\link{conda_remove}()},
\code{\link{conda_version}()}
}
\concept{conda-tools}
