% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_list}
\alias{conda_list}
\title{List Conda Environments}
\usage{
conda_list(conda = "auto")
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}
}
\value{
An \R \code{data.frame}, with \code{name} giving the name of the associated
environment, and \code{python} giving the path to the Python binary associated
with that environment.
}
\description{
List all of the available conda environments on the system.
}
\details{
Environments available are listed as by:\preformatted{conda info --json
}

Note that this function will report \emph{all} available conda environments
on the system, even those associated with a different conda installation
than the requested \code{conda} binary.
}
\seealso{
Other conda-tools: 
\code{\link{conda_binary}()},
\code{\link{conda_create}()},
\code{\link{conda_export}()},
\code{\link{conda_install}()},
\code{\link{conda_python}()},
\code{\link{conda_remove}()},
\code{\link{conda_version}()}
}
\concept{conda-tools}
