% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_remove}
\alias{conda_remove}
\title{Remove Packages from a Conda Environment}
\usage{
conda_remove(envname, packages = NULL, conda = "auto")
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{packages}{An optional character vector, giving the names of packages
to be removed from the conda environment. When \code{NULL} (the default),
everything in the associated environment will be removed.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}
}
\description{
Use this function to remove some subset of packages from an existing
conda environment, or (when \code{packages} is \code{NULL}) remove all packages
from the requested conda environment.
}
\seealso{
Other conda tools: 
\code{\link{conda_binary}()},
\code{\link{conda_create}()},
\code{\link{conda_export}()},
\code{\link{conda_install}()},
\code{\link{conda_list}()},
\code{\link{conda_python}()},
\code{\link{conda_version}()}
}
\concept{conda tools}
