% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_update}
\alias{conda_update}
\title{Update Conda}
\usage{
conda_update(conda = "auto")
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}
}
\description{
Update to the latest-available version of \code{conda}. Equivalent to calling:
}
\details{
\preformatted{conda update --prefix $(conda info --base) conda
}
}
