% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_version}
\alias{conda_version}
\title{Retrieve the Conda Version}
\usage{
conda_version(conda = "auto")
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}
}
\description{
Retrieve the current version of conda, as reported by \code{conda --version}.
}
\seealso{
Other conda-tools: 
\code{\link{conda_binary}()},
\code{\link{conda_create}()},
\code{\link{conda_export}()},
\code{\link{conda_install}()},
\code{\link{conda_list}()},
\code{\link{conda_python}()},
\code{\link{conda_remove}()}
}
\concept{conda-tools}
