% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{constant}
\alias{constant}
\title{Create a constant tensor.}
\usage{
constant(value, dtype = NULL, shape = NULL, name = "Const")
}
\arguments{
\item{value}{A constant value (or list) of output type \code{dtype}.}

\item{dtype}{The type of the elements of the resulting tensor.}

\item{shape}{Optional dimensions of resulting tensor.}

\item{name}{Optional name for the tensor.}
}
\value{
A Constant Tensor.
}
\description{
The resulting tensor is populated with values of type \code{dtype}, as
specified by arguments \code{value} and (optionally) \code{shape} (see
examples below).
}
\details{
The argument \code{value} can be a constant value, or a list of values of
type \code{dtype}. If \code{value} is a list, then the length of the list
must be less than or equal to the number of elements implied by the
\code{shape} argument (if specified). In the case where the list length is
less than the number of elements specified by \code{shape}, the last element
in the list will be used to fill the remaining entries.

The argument \code{shape} is optional. If present, it specifies the
dimensions of the resulting tensor. If not present, the shape of \code{value}
is used.

If the argument \code{dtype} is not specified, then the type is inferred from
the type of \code{value}.
}

