% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python.R
\name{install_python}
\alias{install_python}
\title{Install Python}
\usage{
install_python(version, list = FALSE, force = FALSE)
}
\arguments{
\item{version}{The version of Python to install.}

\item{list}{Boolean; if set, list the set of available Python versions?}

\item{force}{Boolean; force re-installation even if the requested version
of Python is already installed?}
}
\description{
Download and install Python.
}
\details{
On Windows, the official binaries as made available from \url{https://ww.python.org}
are used. On macOS and Linux, Python is built from sources via the
\href{https://github.com/pyenv/pyenv}{pyenv} module.
}
