% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_py_object}
\alias{is_py_object}
\title{Check if x is a Python object}
\usage{
is_py_object(x)
}
\arguments{
\item{x}{An \R or Python.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks if \code{x} is a Python object, more efficiently
than \code{inherits(x, "python.builtin.object")}.
}
\keyword{internal}
