% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{miniconda-params}
\alias{miniconda-params}
\title{miniconda-params}
\arguments{
\item{path}{The location where Miniconda is (or should be) installed. Note
that the Miniconda installer does not support paths containing spaces. See
\link{miniconda_path} for more details on the default path used by \code{reticulate}.}
}
\description{
miniconda-params
}
\keyword{internal}
