% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{miniconda_path}
\alias{miniconda_path}
\title{Path to Miniconda}
\usage{
miniconda_path()
}
\description{
The path to the Miniconda installation to use. By default, an OS-specific
path is used. If you'd like to instead set your own path, you can set the
\code{RETICULATE_MINICONDA_PATH} environment variable.
}
\seealso{
Other miniconda: 
\code{\link{install_miniconda}()},
\code{\link{miniconda_update}()}
}
\concept{miniconda}
