% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{miniconda_uninstall}
\alias{miniconda_uninstall}
\title{Remove Miniconda}
\usage{
miniconda_uninstall(path = miniconda_path())
}
\arguments{
\item{path}{The path in which Miniconda is installed.}
}
\description{
Uninstall Miniconda.
}
\seealso{
Other miniconda-tools: 
\code{\link{install_miniconda}()},
\code{\link{miniconda_update}()}
}
\concept{miniconda-tools}
