% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{miniconda_update}
\alias{miniconda_update}
\title{Update Miniconda}
\usage{
miniconda_update(path = miniconda_path())
}
\arguments{
\item{path}{The location where Miniconda is (or should be) installed. Note
that the Miniconda installer does not support paths containing spaces. See
\link{miniconda_path} for more details on the default path used by \code{reticulate}.}
}
\description{
Update Miniconda to the latest version.
}
\seealso{
Other Miniconda tools: 
\code{\link{install_miniconda}()},
\code{\link{miniconda_uninstall}()}
}
\concept{Miniconda tools}
