% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{nameOfClass.python.builtin.type}
\alias{nameOfClass.python.builtin.type}
\title{\code{nameOfClass()} for Python objects}
\usage{
\method{nameOfClass}{python.builtin.type}(x)
}
\arguments{
\item{x}{A Python class}
}
\value{
A scalar string matching the S3 class of objects constructed from the
type.
}
\description{
This generic enables passing a \code{python.builtin.type} object as the 2nd
argument to \code{base::inherits()}.
}
\examples{
\dontrun{
  numpy <- import("numpy")
  x <- r_to_py(array(1:3))
  inherits(x, numpy$ndarray)
}
}
