% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{np_array}
\alias{np_array}
\title{NumPy array}
\usage{
np_array(data, dtype = NULL, order = "C")
}
\arguments{
\item{data}{Vector or existing NumPy array providing data for the array}

\item{dtype}{Numpy data type (e.g. "float32", "float64", etc.)}

\item{order}{Memory ordering for array. "C" means C order, "F" means Fortran
order.}
}
\value{
A NumPy array object.
}
\description{
Create NumPy arrays and convert the data type and in-memory
ordering of existing NumPy arrays.
}
