% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_flags.R
\name{parse_flags}
\alias{parse_flags}
\title{Parse script FLAGS from command line}
\usage{
parse_flags()
}
\description{
Parse command line arguments to the Rscript and use them to
populate the values of TensorFlow FLAGS
}

