% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_bool}
\alias{py_bool}
\title{Python Truthiness}
\usage{
py_bool(x)
}
\arguments{
\item{x, }{A python object.}
}
\value{
An R scalar logical: \code{TRUE} or \code{FALSE}
}
\description{
Equivalent to \code{bool(x)} in Python, or \verb{non not x}.
}
\details{
If the python objects defines a \verb{__bool__} method, then that is invoked.
Otherwise, if the object defines a \verb{__len__} method, then \code{TRUE} is returned
if the length is nonzero. If neither
\verb{__len__} nor \verb{__bool__} are defined, then the python object is considered
\code{TRUE}.
}
