% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_call}
\alias{py_call}
\title{Call a Python callable object}
\usage{
py_call(x, ...)
}
\arguments{
\item{...}{Arguments to function (named and/or unnamed)}
}
\value{
Return value of call as a Python object.
}
\description{
Call a Python callable object
}
\keyword{internal}
