% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_call_object}
\alias{py_call_object}
\title{Call a Python callable object}
\usage{
py_call_object(x, args, keywords = NULL)
}
\arguments{
\item{args}{List of unnamed arguments}

\item{keywords}{List of named arguments}
}
\value{
Return value of call
}
\description{
Call a Python callable object
}
\keyword{internal}
