% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_capture_output}
\alias{py_capture_output}
\title{Capture and return Python output}
\usage{
py_capture_output(expr, capture = c("all", "stdout", "stderr"))
}
\arguments{
\item{expr}{Expression to capture stdout for}

\item{capture}{Streams to capture}
}
\value{
Character vector with output
}
\description{
Capture and return Python output
}
