% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_config}
\alias{py_config}
\title{Python configuration}
\usage{
py_config()
}
\value{
Information about the version of Python in use, as an \R list with
class \code{"py_config"}.
}
\description{
Retrieve information about the version of Python currently being used by
\code{reticulate}.
}
\details{
If Python has not yet been initialized, then calling \code{py_config()} will force
the initialization of Python. See \code{\link[=py_discover_config]{py_discover_config()}} for more details.
}
