% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_config_error_message}
\alias{py_config_error_message}
\title{Build Python configuration error message}
\usage{
py_config_error_message(prefix)
}
\arguments{
\item{prefix}{Error message prefix}
}
\description{
Build Python configuration error message
}
\keyword{internal}
