% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_del_attr}
\alias{py_del_attr}
\title{Delete an attribute of a Python object}
\usage{
py_del_attr(x, name)
}
\arguments{
\item{x}{A Python object.}

\item{name}{The attribute name.}
}
\description{
Delete an attribute of a Python object
}
