% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_dict}
\alias{py_dict}
\title{Create python dictionary from the lists of keys and values.}
\usage{
py_dict(keys, items, convert)
}
\arguments{
\item{keys}{List of keys which can be python objects.}

\item{items}{List of values.}

\item{convert}{TRUE to automatically convert Python objects to their R equivalent.}
}
\description{
Create python dictionary from the lists of keys and values.
}
