% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_exe}
\alias{py_exe}
\title{Python executable}
\usage{
py_exe()
}
\value{
The path to the associated Python executable, or \code{NULL} if Python
has not yet been initialized.
}
\description{
Get the path to the Python executable associated with the instance currently
being used by \code{reticulate}.
}
\details{
This can occasionally be useful if you'd like to interact with Python (or its
modules) via a subprocess; for example you might choose to install a package
with \code{pip}:\preformatted{system2(py_exe(), c("-m", "pip", "install", "numpy"))
}

and so you can also have greater control over how these modules are invoked.
}
