% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_has_attr}
\alias{py_has_attr}
\title{Check if a Python object has an attribute}
\usage{
py_has_attr(x, name)
}
\arguments{
\item{x}{Python object}

\item{name}{Attribute name}
}
\value{
Logical indicating whether it has the specified attribute
}
\description{
Check if a Python object has an attribute
}
