% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{py_help_handler}
\alias{py_help_handler}
\title{Provide help for Python objects}
\usage{
py_help_handler(type = c("completion", "parameter", "url"), topic, source, ...)
}
\description{
This is an internal method to be used by front-ends which need to provide
help text / information for Python objects in different contexts.
}
\keyword{internal}
