% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_import}
\alias{py_import}
\title{Obtain a reference to a python module}
\usage{
py_import(module)
}
\arguments{
\item{module}{Name of module}
}
\description{
Obtain a reference to a python module
}

