% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_len}
\alias{py_len}
\title{Length of Python object}
\usage{
py_len(x)
}
\arguments{
\item{Python}{object}
}
\value{
Length as integer
}
\description{
Get the length of a Python object (equivalent to the Python \code{len()}
built in function).
}
