% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_list_attributes}
\alias{py_list_attributes}
\title{List all attributes of a Python object}
\usage{
py_list_attributes(x)
}
\arguments{
\item{x}{Python object}
}
\value{
Character vector of attributes
}
\description{
List all attributes of a Python object
}
