% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_module}
\alias{py_module}
\title{Import a Python module}
\usage{
py_module(module = "__main__", silent = FALSE)
}
\arguments{
\item{module}{Module name}

\item{silent}{Return \code{NULL} rather than throwing an error
if the specified module cannot be loaded.}
}
\value{
A Python module
}
\description{
Import the specified Python module for calling from R. If no module
name is specified then the \code{__main__} module is imported.
}

