% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_module_available}
\alias{py_module_available}
\title{Check if a Python module is available on this system.}
\usage{
py_module_available(module)
}
\arguments{
\item{module}{Name of module}
}
\value{
Logical indicating whether module is available
}
\description{
Check if a Python module is available on this system.
}
