% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_register_load_hook}
\alias{py_register_load_hook}
\title{Register a Python module load hook}
\usage{
py_register_load_hook(module, hook)
}
\arguments{
\item{module}{String, the name of the Python module.}

\item{hook}{Function, called with no arguments. If \code{module} is already
loaded, \code{hook()} is called immediately.}
}
\value{
\code{NULL} invisibly. Called for its side effect.
}
\description{
Register an R function to be called when a Python module is first loaded in
the current R session. This can be used for tasks such as:
}
\details{
\itemize{
\item Delayed registration of S3 methods to accommodate different versions of a Python module.
\item Configuring module-specific logging streams.
}
}
\keyword{internal}
