% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{py_repl}
\alias{py_repl}
\title{Run a Python REPL}
\usage{
py_repl(module = getOption("reticulate.repl.module", default = character()),
  quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean; print a startup banner when launching the REPL? If
\code{FALSE}, the banner will be suppressed.}

\item{modules}{An (optional) Python module to be imported before
the REPL is launched. Defaults to the value of the \code{reticulate.repl.module}
option.}
}
\description{
This function provides a Python REPL in the R session, which can be used
to interactively run Python code. All code executed within the REPL is
run within the Python 'main' module, and will remain accessible after the
REPL is detached.
}
