% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{py_repl_active}
\alias{py_repl_active}
\title{Check Whether the Python REPL is Active}
\usage{
py_repl_active()
}
\description{
Check to see whether the Python REPL is active. This is primarily
for use by R front-ends, which might want to toggle or affect
the state of the Python REPL while it is running.
}
