% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_set_attr}
\alias{py_set_attr}
\title{Set an attribute of a Python object}
\usage{
py_set_attr(x, name, value)
}
\arguments{
\item{x}{Python object}

\item{name}{Attribute name}

\item{value}{Attribute value}
}
\description{
Set an attribute of a Python object
}
