% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_slice}
\alias{py_slice}
\title{Return a python \code{slice} object}
\usage{
py_slice(start = NULL, stop = NULL, step = NULL)
}
\arguments{
\item{start, stop, step}{An integer, or NULL.}
}
\value{
a python slice object
}
\description{
Return a python \code{slice} object
}
