% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_suppress_warnings}
\alias{py_suppress_warnings}
\title{Suppress Python warnings for an expression}
\usage{
py_suppress_warnings(expr)
}
\arguments{
\item{expr}{Expression to suppress warnings for}
}
\value{
Result of evaluating expression
}
\description{
Suppress Python warnings for an expression
}
