% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{py_to_r.pandas.core.frame.DataFrame}
\alias{py_to_r.pandas.core.frame.DataFrame}
\title{py_to_r.pandas.core.frame.DataFrame}
\usage{
\method{py_to_r}{pandas.core.frame.DataFrame}(x)
}
\arguments{
\item{x}{A Python object.}
}
