% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_to_r}
\alias{py_to_r}
\title{Convert a Python object to an R object}
\usage{
py_to_r(x)
}
\arguments{
\item{x}{Python object to convert}
}
\value{
R object
}
\description{
Convert a Python object to an R object
}
