% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{py_to_r_wrapper}
\alias{py_to_r_wrapper}
\title{R wrapper for Python objects}
\usage{
py_to_r_wrapper(x)
}
\arguments{
\item{x}{Python object}
}
\description{
S3 method to create a custom R wrapper for a Python object.
The default wrapper is either an R environment or an R function
(for callable python objects).
}
\keyword{internal}
