% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_unicode}
\alias{py_unicode}
\title{Convert to Python Unicode Object}
\usage{
py_unicode(str)
}
\arguments{
\item{str}{Single element character vector to convert}
}
\description{
Convert to Python Unicode Object
}
\details{
By default R character vectors are converted to Python strings.
In Python 3 these values are unicode objects however in Python 2
they are 8-bit string objects. This function enables you to
obtain a Python unicode object from an R character vector
when running under Python 2 (under Python 3 a standard Python
string object is returned).
}
