% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_versions_windows}
\alias{py_versions_windows}
\title{Discover versions of Python installed on a Windows system}
\usage{
py_versions_windows()
}
\value{
Data frame with \code{type}, \code{hive}, \code{install_path}, \code{executable_path},
and \code{version}.
}
\description{
Discover versions of Python installed on a Windows system
}
\keyword{internal}
