% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_xptr_str}
\alias{py_xptr_str}
\title{Evaluate an expression that prints a string with
a check for a null externalptr}
\usage{
py_xptr_str(x, expr)
}
\arguments{
\item{x}{Python object}

\item{expr}{R expression that uses the object
to print a string.}
}
\description{
Evaluate an expression that prints a string with
a check for a null externalptr
}
\details{
If the object is a null externalptr
then \code{<pointer: 0x0>} will be printed.
Otherwise the expression will be evaluated.
}
