% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\docType{data}
\name{python-aliases}
\alias{False}
\alias{None}
\alias{True}
\alias{python-aliases}
\title{R aliases for various Python literals}
\format{An R object semantically equivalent to the Python literal.}
\usage{
None

True

False
}
\description{
Aliases for various Python literals which make it easier to port existing
TensorFlow Python code to R (note that the R equivalent R literals e.g.
\code{NULL}, \code{TRUE}, etc. can always still be used).
}
\details{
\code{None} is an alias for \code{NULL}.

\code{True} is an alias for \code{TRUE}.

\code{False} is an alias for \code{FALSE}.
}
\keyword{datasets}

