% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{python_repl}
\alias{python_repl}
\title{Run a Python REPL}
\usage{
python_repl(banner = TRUE)
}
\arguments{
\item{banner}{Boolean; print a startup banner when launching the REPL?}
}
\description{
This function provides a Python REPL in the R session, which can be used
to interactively run Python code. All code executed within the REPL is
run within the Python 'main' module, and will remain accessible after the
REPL is detached.
}
