% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{register_class_filter}
\alias{register_class_filter}
\title{Register a filter for class names}
\usage{
register_class_filter(filter)
}
\arguments{
\item{filter}{Function which takes a class name and maps it to an alternate
name}
}
\description{
Register a filter for class names
}
\keyword{internal}
