% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{register_help_topics}
\alias{register_help_topics}
\title{Register help topics}
\usage{
register_help_topics(type = c("module", "class"), topics)
}
\arguments{
\item{type}{Type (module or class)}

\item{topics}{Named list of topics}
}
\description{
Register a set of help topics for dispatching from F1 help
}
\keyword{internal}
