% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{register_module_help_handler}
\alias{register_module_help_handler}
\title{Register a help handler for a root Python module}
\usage{
register_module_help_handler(module, handler)
}
\arguments{
\item{module}{Name of a root Python module}

\item{handler}{Handler function: \verb{function(name, subtopic = NULL)}. The name
will be the fully qualified name of a Python object (module, function, or
class). The \code{subtopic} is optional and is currently used only for methods
within classes.}
}
\description{
Register a help handler for a root Python module
}
\details{
The help handler is passed a fully qualified module, class, or
function name (and optional method for classes). It should return a URL for
a help page (or \code{""} if no help page is available).
}
\keyword{internal}
