% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{register_suppress_warnings_handler}
\alias{register_suppress_warnings_handler}
\title{Register a handler for calls to py_suppress_warnings}
\usage{
register_suppress_warnings_handler(handler)
}
\arguments{
\item{handler}{Handler}
}
\description{
Register a handler for calls to py_suppress_warnings
}
\details{
Enables packages to register a pair of functions
to be called to suppress and then re-enable warnings
}
\keyword{internal}
