% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\title{Create a new TensorFlow session}
\usage{
session(target = "", graph = NULL, config = NULL)
}
\arguments{
\item{target}{Optional. The execution engine to connect to. Defaults to using
an in-process engine.}

\item{graph}{Optional. The Graph to be launched (described below).}

\item{config}{(Optional.) A list with configuration options for the session.}
}
\description{
A Session object encapsulates the environment in which Operation objects are
executed, and Tensor objects are evaluated.
}
\details{
If no \code{graph} argument is specified when constructing the session, the
default graph will be launched in the session. If you are using more than one
graph (created with \code{graph}) in the same process, you will have to use
different sessions for each graph, but each graph can be used in multiple
sessions. In this case, it is often clearer to pass the graph to be launched
explicitly to the session function
}

