% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Suppress Python and TensorFlow warnings for an expression}
\usage{
suppress_warnings(expr)
}
\arguments{
\item{expr}{Expression to suppress warnigns for}
}
\value{
Result of evaluating expression
}
\description{
Suppress Python and TensorFlow warnings for an expression
}

