% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tensorflow}
\alias{tensorflow}
\alias{tensorflow-package}
\title{TensorFlow for R}
\description{
\href{https://tensorflow.org}{TensorFlow} is an open source software library
for numerical computation using data flow graphs. Nodes in the graph
represent mathematical operations, while the graph edges represent the
multidimensional data arrays (tensors) communicated between them. The
flexible architecture allows you to deploy computation to one or more CPUs or
GPUs in a desktop, server, or mobile device with a single API.
}
\details{
The \href{https://www.tensorflow.org/api_docs/python/index.html}{TensorFlow
API} is composed of a set of Python modules that enable constructing and
executing TensorFlow graphs. The tensorflow package provides access to the
complete TensorFlow API from within R.

For additional documentation on the tensorflow package see
\href{https://rstudio.github.io/tensorflow}{https://rstudio.github.io/tensorflow}
}

