% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.bfloat16}
\alias{tf.bfloat16}
\alias{tf.bfloat16_ref}
\title{TensorFlow 16-bit truncated floating-point.}
\format{TensorFlow DType}
\usage{
tf.bfloat16

tf.bfloat16_ref
}
\description{
TensorFlow 16-bit truncated floating-point.
}
\keyword{datasets}

