% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.complex128}
\alias{tf.complex128}
\alias{tf.complex128_ref}
\title{TensorFlow 128-bit double-precision complex.}
\format{TensorFlow DType}
\usage{
tf.complex128

tf.complex128_ref
}
\description{
TensorFlow 128-bit double-precision complex.
}
\keyword{datasets}

