% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.complex64}
\alias{tf.complex64}
\alias{tf.complex64_ref}
\title{TensorFlow 64-bit single-precision complex.}
\format{TensorFlow DType}
\usage{
tf.complex64

tf.complex64_ref
}
\description{
TensorFlow 64-bit single-precision complex.
}
\keyword{datasets}

