% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.float16}
\alias{tf.float16}
\alias{tf.float16_ref}
\alias{tf.float32_ref}
\alias{tf.float64_ref}
\title{TensorFlow 16-bit half-precision floating-point}
\format{TensorFlow DType}
\usage{
tf.float16

tf.float16_ref

tf.float32_ref

tf.float64_ref
}
\description{
TensorFlow 16-bit half-precision floating-point
}
\keyword{datasets}

