% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.float32}
\alias{tf.float32}
\title{TensorFlow 32-bit single-precision floating-point.}
\format{TensorFlow DType}
\usage{
tf.float32
}
\description{
TensorFlow 32-bit single-precision floating-point.
}
\keyword{datasets}

