% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.float64}
\alias{tf.float64}
\title{TensorFlow 64-bit single-precision floating-point.}
\format{TensorFlow DType}
\usage{
tf.float64
}
\description{
TensorFlow 64-bit single-precision floating-point.
}
\keyword{datasets}

