% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.int16}
\alias{tf.int16}
\alias{tf.int16_ref}
\title{TensorFlow 16-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.int16

tf.int16_ref
}
\description{
TensorFlow 16-bit signed integer.
}
\keyword{datasets}

