% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.int32}
\alias{tf.int32}
\alias{tf.int32_ref}
\title{TensorFlow 32-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.int32

tf.int32_ref
}
\description{
TensorFlow 32-bit signed integer.
}
\keyword{datasets}

