% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.int64}
\alias{tf.int64}
\alias{tf.int64_ref}
\title{TensorFlow 64-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.int64

tf.int64_ref
}
\description{
TensorFlow 64-bit signed integer.
}
\keyword{datasets}

