% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.int8}
\alias{tf.int8}
\alias{tf.int8_ref}
\title{TensorFlow 8-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.int8

tf.int8_ref
}
\description{
TensorFlow 8-bit signed integer.
}
\keyword{datasets}

