% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.qint16}
\alias{tf.qint16}
\alias{tf.qint16_ref}
\title{TensorFlow quantized 16-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.qint16

tf.qint16_ref
}
\description{
TensorFlow quantized 16-bit signed integer.
}
\keyword{datasets}

