% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.qint32}
\alias{tf.qint32}
\alias{tf.qint32_ref}
\title{TensorFlow quantized 32-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.qint32

tf.qint32_ref
}
\description{
TensorFlow quantized 32-bit signed integer.
}
\keyword{datasets}

