% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.qint8}
\alias{tf.qint8}
\alias{tf.qint8_ref}
\title{TensorFlow quantized 8-bit signed integer.}
\format{TensorFlow DType}
\usage{
tf.qint8

tf.qint8_ref
}
\description{
TensorFlow quantized 8-bit signed integer.
}
\keyword{datasets}

