% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.quint16}
\alias{tf.quint16}
\alias{tf.quint16_ref}
\title{TensorFlow quantized 16-bit unsigned integer.}
\format{TensorFlow DType}
\usage{
tf.quint16

tf.quint16_ref
}
\description{
TensorFlow quantized 16-bit unsigned integer.
}
\keyword{datasets}

