% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.quint8}
\alias{tf.quint8}
\alias{tf.quint8_ref}
\title{TensorFlow quantized 8-bit unsigned integer.}
\format{TensorFlow DType}
\usage{
tf.quint8

tf.quint8_ref
}
\description{
TensorFlow quantized 8-bit unsigned integer.
}
\keyword{datasets}

