% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_modules.R
\docType{data}
\name{tf}
\alias{tf}
\title{Main TensorFlow module}
\format{TensorFlow module}
\usage{
tf
}
\description{
Interface to main TensorFlow  module. Provides access to top level classes
and functions as well as sub-modules (e.g. \code{tf$nn},
\code{tf$contrib$learn}, etc.).
}
\examples{
\dontrun{
hello <- tf$constant('Hello, TensorFlow!')
b <- tf$Variable(tf$zeros(shape(1L)))

sess <- tf$Session()
sess$initialize_all_variables()

learn <- tf$contrib$learn
slim <- tf$contrib$slim
}
}
\keyword{datasets}

