% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.uint16}
\alias{tf.uint16}
\alias{tf.uint16_ref}
\title{TensorFlow 16-bit unsigned integer.}
\format{TensorFlow DType}
\usage{
tf.uint16

tf.uint16_ref
}
\description{
TensorFlow 16-bit unsigned integer.
}
\keyword{datasets}

