% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\docType{data}
\name{tf.uint8}
\alias{tf.uint8}
\alias{tf.uint8_ref}
\title{TensorFlow 8-bit unsigned integer.}
\format{TensorFlow DType}
\usage{
tf.uint8

tf.uint8_ref
}
\description{
TensorFlow 8-bit unsigned integer.
}
\keyword{datasets}

