% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_config.R
\name{tf_config}
\alias{tf_config}
\title{TensorFlow configuration information}
\usage{
tf_config()
}
\description{
Retreive a list containing configuration information including the
\code{python} binary and \code{libpython} the package was built against,
, and the version and location of the TensorFlow python module.
}

