% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{tf_import}
\alias{tf_import}
\title{Obtain a reference to a tensorflow module}
\usage{
tf_import(module = "tensorflow")
}
\arguments{
\item{module}{Name of module (defaults to "tensorflow")}
}
\description{
Obtain a reference to a tensorflow module
}
\examples{
\dontrun{
tf <- tf_import()
input_data <- tf_import("examples.tutorials.mnist.input_data")
}
}

