% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{use_python}
\alias{use_python}
\alias{use_virtualenv}
\alias{use_condaenv}
\alias{use_miniconda}
\alias{use_python_version}
\title{Use Python}
\usage{
use_python(python, required = FALSE)

use_virtualenv(virtualenv = NULL, required = FALSE)

use_condaenv(condaenv = NULL, conda = "auto", required = FALSE)

use_miniconda(condaenv = NULL, required = FALSE)

use_python_version(version, required = FALSE)
}
\arguments{
\item{python}{The path to a Python binary.}

\item{required}{Is the requested copy of Python required? If \code{TRUE}, an error will be
emitted if the requested copy of Python does not exist. Otherwise, the
request is taken as a hint only, and scanning for other versions will still
proceed.}

\item{virtualenv}{Either the name of, or the path to, a Python virtual environment.}

\item{condaenv}{Then name of the Conda environment to use.}

\item{conda}{The path to a \code{conda} executable. By default, \code{reticulate} will check the
\code{PATH}, as well as other standard locations for Anaconda installations.}

\item{version}{The version of Python to use. \code{reticulate} will search for versions of
Python as installed by the \code{\link[=install_python]{install_python()}} helper function.}
}
\description{
Control which copy of Python \code{reticulate} will use when it initializes
its Python bindings.
}
\details{
The \code{RETICULATE_PYTHON} environment variable can also be used to control
which copy of Python \code{reticulate} chooses to bind to. When set, this will
silently override any other requests to use a particular copy of Python.
}
