% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_fn.R
\name{wrap_fn}
\alias{wrap_fn}
\title{Wrap R function into Python function with the same signature.}
\usage{
wrap_fn(f)
}
\arguments{
\item{f}{An R function}
}
\value{
A Python function with the same signature as \code{f}.
}
\description{
This function could wrap an R function into Python function with
the same signature. Note that this is function still experimental.
}
