% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-cache.R
\name{cache_eng_python}
\alias{cache_eng_python}
\title{A reticulate cache engine for Knitr}
\usage{
cache_eng_python(options)
}
\arguments{
\item{options}{List of chunk options provided by \code{knitr} during chunk execution.
Contains the caching path.}
}
\description{
This provides caching of Python variables to the \code{reticulate} engine for
\code{knitr}. The cache allows \code{knitr} to save and load the state of Python
variables between cached chunks.  The cache engine depends on the \code{dill}
Python module. Therefore, you must have a recent version of \code{dill} installed
in your Python environment.
}
\details{
The Python cache is activated the same way as the R cache, by setting the
\code{cache} chunk option to \code{TRUE}.  To \emph{deactivate} the Python cache globally
while keeping the R cache active, one may set the option \code{reticulate.cache}
to \code{FALSE}.  For example:

\preformatted{knitr::opts_knit$set(reticulate.cache = FALSE)}
}
\note{
Different from \code{knitr}'s R cache, the Python cache is capable of saving
most, but not all types of Python objects.  Some Python objects are
"unpickleable" and will rise an error when attepmted to be saved.
}
