% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{conda-tools}
\alias{conda-tools}
\alias{conda_list}
\alias{conda_create}
\alias{conda_install}
\title{Interface to conda utility commands}
\usage{
conda_list(conda = "auto")

conda_create(envname, conda = "auto")

conda_install(envname, pkgs, pip = FALSE, conda = "auto")
}
\arguments{
\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations).}

\item{envname}{Name of conda environment}

\item{pkgs}{Character vector with package names to install.}

\item{pip}{\code{TRUE} to use pip (defaults to \code{FALSE})}
}
\value{
\code{conda_list()} returns a data frame with the names and paths to the
respective python binaries of available environments. \code{conda_create()}
returns the Path to the python binary of the created environment.
}
\description{
Interface to conda utility commands
}
\keyword{intername}
