% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{conda-tools}
\alias{conda-tools}
\alias{conda_list}
\alias{conda_create}
\alias{conda_remove}
\alias{conda_install}
\alias{conda_binary}
\title{Interface to conda utility commands}
\usage{
conda_list(conda = "auto")

conda_create(envname, packages = "python", conda = "auto")

conda_remove(envname, packages = NULL, conda = "auto")

conda_install(envname, packages, pip = FALSE, conda = "auto")

conda_binary(conda = "auto")
}
\arguments{
\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations).}

\item{envname}{Name of conda environment}

\item{pip}{\code{TRUE} to use pip (defaults to \code{FALSE})}

\item{pkgs}{Character vector with package names to install.}
}
\value{
\code{conda_list()} returns a data frame with the names and paths to the
respective python binaries of available environments. \code{conda_create()}
returns the Path to the python binary of the created environment.
\code{conda_binary()} returns the location of the conda binary or \code{NULL}
if none can be found.

\code{conda_create()} can be called with the name of an existing
environment (in that case it simply returns the path to the python
binary within that environment).
}
\description{
Interface to conda utility commands
}
\keyword{intername}
