% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_python}
\alias{conda_python}
\title{Get Python Path in Conda Environment}
\usage{
conda_python(envname = NULL, conda = "auto", all = FALSE)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\code{\link[=conda_binary]{conda_binary()}} for more details on how \code{reticulate} tries to resolve
the \code{conda} executable.}

\item{all}{Boolean; should all conda environments with the requested name
be returned? By default, only the first matching environment is returned.}
}
\description{
Find the path to the \code{python} executable associated with a particular
conda environment.
}
\details{
Use \code{envname = "base"} to request the "base" environment associated with
a Conda installation.
}
\seealso{
Other conda-tools: 
\code{\link{conda_binary}()},
\code{\link{conda_create}()},
\code{\link{conda_export}()},
\code{\link{conda_install}()},
\code{\link{conda_list}()},
\code{\link{conda_remove}()},
\code{\link{conda_version}()}
}
\concept{conda-tools}
