% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-packages.R
\name{configure_environment}
\alias{configure_environment}
\title{Configure a Python Environment}
\usage{
configure_environment(package = NULL)
}
\description{
Configure a Python environment, satisfying the Python dependencies of any
loaded \R packages.
}
\details{
Normally, this function should only be used by package authors, who want
to ensure that their package dependencies are installed in the active
Python environment. For example:\preformatted{.onLoad <- function(libname, pkgname) \{
  reticulate::configure_environment()
\}
}

If the Python session has not yet been initialized, or if the user is not
using the default Miniconda Python installation, no action will be taken.
Otherwise, \code{reticulate} will take this as a signal to install any required
Python dependencies into the user's Python environment.

Note that, in the case where the Python session has not yet been initialized,
\code{reticulate} will automatically ensure your required Python dependencies
are installed after the Python session is initialized (when appropriate).
}
