% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{dict}
\alias{dict}
\alias{py_dict}
\title{Create Python dictionary}
\usage{
dict(..., convert = FALSE)

py_dict(keys, values, convert = FALSE)
}
\arguments{
\item{...}{Name/value pairs for dictionary (or a single named list to be
converted to a dictionary).}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R
equivalent. If you pass \code{FALSE} you can do manual conversion using the
\code{\link[=py_to_r]{py_to_r()}} function.}

\item{keys}{Keys to dictionary (can be Python objects)}

\item{values}{Values for dictionary}
}
\value{
A Python dictionary
}
\description{
Create a Python dictionary object, including a dictionary whose keys are
other Python objects rather than character vectors.
}
\note{
The returned dictionary will not automatically convert its elements
from Python to R. You can do manual converstion with the \code{\link[=py_to_r]{py_to_r()}}
function or pass \code{convert = TRUE} to request automatic conversion.
}
