% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{dict}
\alias{dict}
\title{Create Python dictionary}
\usage{
dict(...)
}
\arguments{
\item{...}{Name/value pairs for dictionary}
}
\value{
A Python dictionary
}
\description{
Create a Python dictionary object, including a dictionary whose keys
are other Python objects rather than character vectors.
}
\note{
This is useful for creating dictionaries keyed by Tensor (required
for `\code{feed_dict}` parameters).
}
