% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{eng_python}
\alias{eng_python}
\title{A reticulate Engine for Knitr}
\usage{
eng_python(options)
}
\arguments{
\item{options}{Chunk options, as provided by \code{knitr} during chunk execution.}
}
\description{
This provides a \code{reticulate} engine for \code{knitr}, suitable for usage when
attempting to render Python chunks. Using this engine allows for shared state
between Python chunks in a document -- that is, variables defined by one
Python chunk can be used by later Python chunks.
}
\details{
The engine can be activated by setting (for example)\preformatted{knitr::knit_engines$set(python = reticulate::eng_python)
}

Typically, this will be set within a document's setup chunk, or by the
environment requesting that Python chunks be processed by this engine.
}
