% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{import}
\alias{import}
\title{Import a Python module}
\usage{
import(module, delay_load = FALSE)
}
\arguments{
\item{module}{Module name}

\item{delay_load}{\code{TRUE} or a function to delay loading the module until
it is first used (if a function is provided then it will be called
once the module is loaded). \code{FALSE} to load the module immediately.}
}
\value{
A Python module
}
\description{
Import the specified Python module for calling from R. Use \code{"__main__"}
to import the main module.
}
\examples{
\dontrun{
main <- import("__main__")
sys <- import("sys")
}

}
