% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{import}
\alias{import}
\alias{import_main}
\alias{import_builtins}
\alias{import_from_path}
\title{Import a Python module}
\usage{
import(module, convert = TRUE, delay_load = FALSE)

import_main(convert = TRUE)

import_builtins(convert = TRUE)

import_from_path(module, path, convert = TRUE, delay_load = FALSE)
}
\arguments{
\item{module}{Module name}

\item{convert}{\code{TRUE} to automatically convert Python objects to their
R equivalent. If you pass \code{FALSE} you can do manual conversion using the
\code{\link[=py_to_r]{py_to_r()}} function.}

\item{delay_load}{\code{TRUE} or a function to delay loading the module until
it is first used (if a function is provided then it will be called
once the module is loaded). \code{FALSE} to load the module immediately.}

\item{package}{Package name}
}
\value{
A Python module
}
\description{
Import the specified Python module for calling from R.
}
\details{
The \code{import_from_package} function imports a Python module
defined within an R package (the module must be located within the
\code{inst/python} directory of the R package).
}
\examples{
\dontrun{
main <- import_main()
sys <- import("sys")
}

}
