% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{install_miniconda}
\alias{install_miniconda}
\title{Install Miniconda}
\usage{
install_miniconda(path = miniconda_path(),
  version = Sys.getenv("RETICULATE_MINICONDA_PYTHON_VERSION", unset =
  "3"), update = TRUE, packages = c("numpy", "scipy", "pandas"),
  force = FALSE)
}
\arguments{
\item{path}{The path in which Miniconda will be installed. Note that the
installer does not support paths containing spaces.}

\item{version}{The (major) version of Python to use.}

\item{update}{Boolean; update to the latest version of Conda after install?}

\item{packages}{A vector of Python packages to install into the Miniconda
installation.}

\item{force}{Boolean; force re-installation if Miniconda is already installed
at the requested path?}
}
\description{
Download the \href{https://docs.conda.io/en/latest/miniconda.html}{Miniconda}
installer, and use it to install Miniconda.
}
\seealso{
Other miniconda: \code{\link{miniconda_path}},
  \code{\link{miniconda_update}}
}
\concept{miniconda}
