% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{iterate}
\alias{iterate}
\title{Traverse a Python iterator or generator}
\usage{
iterate(x, f = base::identity)
}
\arguments{
\item{x}{Python iterator or generator}

\item{f}{Function to apply to each item. By default applies
the \code{identity} function which just reflects back the
value of the item.}
}
\value{
List containing the results of calling \code{f} on each
 item in \code{x} (invisibly).
}
\description{
Traverse a Python iterator or generator
}

