% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{miniconda_path}
\alias{miniconda_path}
\title{Path to Miniconda}
\usage{
miniconda_path()
}
\description{
The path to the Miniconda installation to use.
}
\seealso{
Other miniconda: \code{\link{install_miniconda}()},
  \code{\link{miniconda_update}()}
}
\concept{miniconda}
