% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniconda.R
\name{miniconda_update}
\alias{miniconda_update}
\title{Update Miniconda}
\usage{
miniconda_update(path = miniconda_path())
}
\arguments{
\item{path}{The path in which Miniconda will be installed. Note that the
installer does not support paths containing spaces.}
}
\description{
Update Miniconda to the latest version.
}
\seealso{
Other miniconda: \code{\link{install_miniconda}()},
  \code{\link{miniconda_path}()}
}
\concept{miniconda}
