% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{py}
\alias{py}
\title{Interact with the Python Main Module}
\format{An \R object acting as an interface to the
Python main module.}
\usage{
py
}
\description{
The \code{py} object provides a means for interacting
with the Python main session directly from \R.
}
\keyword{datasets}
