% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_available}
\alias{py_available}
\title{Check if Python is available on this system}
\usage{
py_available(initialize = TRUE)
}
\arguments{
\item{initialize}{\code{TRUE} to attempt to initialize Python bindings if they
aren't yet available.}
}
\value{
Logical indicating whether Python is initialized.
}
\description{
Check if Python is available on this system
}
