% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_call}
\alias{py_call}
\title{Call a Python callable object}
\usage{
py_call(x, args = NULL, keywords = NULL, convert = TRUE)
}
\arguments{
\item{args}{List of unnamed arguments}

\item{keywords}{List of named arguments}

\item{convert}{\code{TRUE} to convert the return value to R (otherwise keep it as
a Python object).}
}
\value{
Return value of call
}
\description{
Call a Python callable object
}
\keyword{internal}
