% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_config.R
\name{py_config}
\alias{py_config}
\title{Python configuration}
\usage{
py_config()
}
\value{
Python configuration object; Logical indicating whether Python
  bindings are available
}
\description{
Information on Python and Numpy versions detected
}
