% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_config}
\alias{py_config}
\title{Python configuration}
\usage{
py_config()
}
\value{
Python configuration object; Logical indicating whether Python
bindings are available
}
\description{
Information on Python and Numpy versions detected.
Note: Python will be initialized and therefore \code{\link[=py_available]{py_available()}} will return
\code{TRUE}. If this is not desired, check the \code{\link[=py_discover_config]{py_discover_config()}} function.
}
