% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_exe}
\alias{py_exe}
\title{Python executable}
\usage{
py_exe()
}
\value{
The path to the Python executable \code{reticulate} has been configured
to use.
}
\description{
Get the path to the Python executable that \code{reticulate} has been configured
to use. If Python has already been initialized, then \code{reticulate} will
choose the currently-active copy of Python.
}
\details{
This can occasionally be useful if you'd like to interact with Python (or its
modules) via a subprocess; for example you might choose to install a package
with \code{pip}:\preformatted{system2(py_exe(), c("-m", "pip", "install", "numpy"))
}

and so you can also have greater control over how these modules are invoked.
}
