% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_has_attr}
\alias{py_has_attr}
\title{Check if a Python object has an attribute}
\usage{
py_has_attr(x, name)
}
\arguments{
\item{x}{A python object.}

\item{name}{The attribute to be accessed.}
}
\value{
\code{TRUE} if the object has the attribute \code{name}, and
\code{FALSE} otherwise.
}
\description{
Check whether a Python object \code{x} has an attribute
\code{name}.
}
