% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_id}
\alias{py_id}
\title{Unique identifer for Python object}
\usage{
py_id(object)
}
\arguments{
\item{object}{Python object}
}
\value{
Unique identifer (as string) or \code{NULL}
}
\description{
Get a globally unique identifier for a Python object.
}
\note{
In the current implementation of CPython this is the
memory address of the object.
}
