% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{py_install}
\alias{py_install}
\title{Install Python packages}
\usage{
py_install(packages, envname = "r-reticulate", method = c("auto",
  "virtualenv", "conda"), conda = "auto")
}
\arguments{
\item{packages}{Character vector with package names to install}

\item{envname}{Name of environment to install packages into}

\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}
}
\description{
Install Python packages into a virtualenv or conda env.
}
\details{
On Linux and OS X the "virtualenv" method will be used by default
("conda" will be used if virtualenv isn't available). On Windows, the
"conda" method is always used.
}
\seealso{
\link{conda-tools}, \link{virtualenv-tools}
}
