% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_clear_last_error}
\alias{py_clear_last_error}
\alias{py_last_error}
\title{Get or clear the last Python error encountered}
\usage{
py_clear_last_error()

py_last_error(exception = getOption("reticulate.last_exception"))
}
\arguments{
\item{exception}{A python exception object.}
}
\value{
For \code{py_last_error()}, \code{NULL} if no error has yet been encountered.
Otherwise, a named list with entries
\itemize{
\item \code{"type"}: R string, name of the exception class.
\item \code{"value"}: R string, formatted exception message.
\item \code{"traceback"}: R character vector, the formatted python traceback,
\item \code{"message"}: The full formatted raised exception, as it would be printed in
Python. Includes the traceback, type, and value.
}

The named list has \code{class} \code{"py_error"}, and has a default \code{print} method
that is the equivalent of \code{cat(py_last_error()$message)}.
}
\description{
Get or clear the last Python error encountered
}
