% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_last_error}
\alias{py_last_error}
\title{Get the last Python error encountered}
\usage{
py_last_error()
}
\description{
Return a list with the type, value, and traceback for the last
Python error encountered (can be \code{NULL} if no error has yet been
encountered).
}
