% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_module}
\alias{py_module}
\title{Import a Python module}
\usage{
py_module(module = "__main__")
}
\arguments{
\item{module}{Module name}
}
\value{
A Python module
}
\description{
Import the specified Python module for calling from R. If no module
name is specified then the \code{__main__} module is imported.
}

