% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{py_module_available}
\alias{py_module_available}
\title{Check if a Python module is available on this system.}
\usage{
py_module_available(module)
}
\arguments{
\item{module}{The name of the module.}
}
\value{
\code{TRUE} if the module is available and can be loaded;
\code{FALSE} otherwise.
}
\description{
Note that this function will also attempt to initialize Python
before checking if the requested module is available.
}
