% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_run}
\alias{py_run}
\alias{py_run_file}
\alias{py_run_string}
\title{Run Python code}
\usage{
py_run_string(code)

py_run_file(file)
}
\arguments{
\item{code}{Code to execute}

\item{file}{File to execute}
}
\description{
Execute code within the the \code{__main__} Python module.
}

