% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_run}
\alias{py_run}
\alias{py_run_string}
\alias{py_run_file}
\title{Run Python code}
\usage{
py_run_string(code, convert = TRUE)

py_run_file(file, convert = TRUE)
}
\arguments{
\item{code}{Code to execute}

\item{convert}{\code{TRUE} to automatically convert Python objects to their
R equivalent. If you pass \code{FALSE} you can do manual conversion using the
\code{\link[=py_to_r]{py_to_r()}} function.}

\item{file}{File to execute}
}
\value{
Reference to \code{__main__} Python module.
}
\description{
Execute code within the the \code{__main__} Python module.
}
