% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_run}
\alias{py_run}
\alias{run_file}
\alias{run_string}
\title{Run Python code}
\usage{
run_string(code)

run_file(file)
}
\arguments{
\item{code}{Code to execute}

\item{file}{File to execute}
}
\value{
Reference to \code{__main__} Python module.
}
\description{
Execute code within the the \code{__main__} Python module.
}

