% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickle.R
\name{py_save_object}
\alias{py_save_object}
\alias{py_load_object}
\title{Save and Load Python Objects}
\usage{
py_save_object(object, filename, pickle = "pickle", ...)

py_load_object(filename, pickle = "pickle", ...)
}
\arguments{
\item{object}{A Python object.}

\item{filename}{The output file name. Note that the file extension \code{.pickle}
is considered the "standard" extension for serialized Python objects
as created by the \code{pickle} module.}

\item{pickle}{The "pickle" implementation to use. Defaults to \verb{"pickle}",
but other compatible Python "pickle" implementations (e.g. \code{"cPickle"})
could be used as well.}

\item{...}{Optional arguments, to be passed to the \code{pickle} module's
\code{dump()} and \code{load()} functions.}
}
\description{
Save and load Python objects.
}
\details{
Python objects are serialized using the \code{pickle} module -- see
\url{https://docs.python.org/3/library/pickle.html} for more details.
}
