% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_set_item}
\alias{py_set_item}
\title{Set an item for a Python object}
\usage{
py_set_item(x, name, value)
}
\arguments{
\item{x}{A Python object.}

\item{name}{The item name.}

\item{value}{The item value.}
}
\value{
The (mutated) object \code{x}, invisibly.
}
\description{
Set an item on a Python object, similar to how
\code{x[name] = value} might be used in Python code to
set an item called \code{name} with value \code{value} on object
\code{x}. The object's \code{__setitem__} method will be called.
}
\seealso{
Other item-related APIs: \code{\link{py_del_item}},
  \code{\link{py_get_item}}
}
\concept{item-related APIs}
