% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_str}
\alias{py_str}
\title{String representation of a python object.}
\usage{
py_str(object, ...)
}
\arguments{
\item{object}{Python object}

\item{...}{Unused}
}
\value{
Character vector
}
\description{
This is equivalent to calling \code{str(object)} in Python, and also the same method
invoked by Python's \code{print()}.
}
\details{
For historical reasons, this is also an \R S3 method that allows R authors to
customize the the string representation of a Python object from R. New code
is recommended to provide a \code{format} S3 R method for python objects instead.

The default implementation will call \code{PyObject_Str} on the object.
}
