% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{py_str}
\alias{py_str}
\title{String representation of a Python object}
\usage{
py_str(x)
}
\arguments{
\item{x}{Python object}
}
\value{
Character vector with result of calling `PyObject_Str` on the object.
}
\description{
String representation of a Python object
}
