% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_str}
\alias{py_str}
\title{An S3 method for getting the string representation of a Python object}
\usage{
py_str(x, ...)
}
\arguments{
\item{x}{Python object}

\item{...}{Unused}
}
\value{
Character vector
}
\description{
An S3 method for getting the string representation of a Python object
}
\details{
The default implementation will call \code{PyObject_Str} on the object.
}
