% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{r-py-conversion}
\alias{r-py-conversion}
\alias{py_to_r}
\title{Convert between Pyton and R objects}
\usage{
py_to_r(x)
}
\arguments{
\item{x}{Object to convert}
}
\value{
Converted object
}
\description{
Convert between Pyton and R objects
}
