% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{r-py-conversion}
\alias{r-py-conversion}
\alias{py_to_r}
\alias{r_to_py}
\title{Convert between Python and R objects}
\usage{
py_to_r(x)

r_to_py(x, convert = FALSE)
}
\arguments{
\item{x}{A Python object.}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R
equivalent. If you pass \code{FALSE} you can do manual conversion using the
\code{\link[=py_to_r]{py_to_r()}} function.}
}
\value{
An \R object, as converted from the Python object.
}
\description{
Convert between Python and R objects
}
