% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{r-py-conversion}
\alias{r-py-conversion}
\alias{r_to_py}
\alias{py_to_r}
\title{Convert between Python and R objects}
\usage{
r_to_py(x, convert = FALSE)

py_to_r(x)
}
\arguments{
\item{x}{A Python object.}
}
\value{
An \R object, as converted from the Python object.
}
\description{
Convert between Python and R objects
}
