% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{reticulate}
\alias{reticulate}
\alias{reticulate-package}
\title{R Interface to Python}
\description{
R interface to Python modules, classes, and functions. When calling into 
Python R data types are automatically converted to their equivalent Python 
types. When values are returned from Python to R they are converted back to R
types. The reticulate package is compatible with all versions of Python >= 2.7 and 
in addition requires NumPy >= 1.11.
}
\details{
Additional documentation can be found at
\href{https://github.com/rstudio/reticulate}{https://github.com/rstudio/reticulate}.
}
