% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorflow.R
\name{tensorflow}
\alias{is_tensorflow_installed}
\alias{tensorflow}
\title{Import TensorFlow}
\usage{
tensorflow(module = NULL, silent = FALSE)

is_tensorflow_installed()
}
\arguments{
\item{module}{Name of sub-module to import. Defaults to \code{NULL}, which
imports the main tensorflow module (\code{tf}).}

\item{silent}{Return \code{NULL} rather than throwing an error
if the specified module cannot be loaded.}
}
\value{
A tensorflow module
}
\description{
Import the tensorflow python module (or one of it's sub-modules) for
use in R.
}
\examples{
\dontrun{
tf <- tensorflow()
learn <- tensorflow("contrib.learn")
slim <- tensorflow("contrib.slim")
}

}

