% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorflow.R
\name{tensorflow}
\alias{tensorflow}
\title{Import TensorFlow}
\usage{
tensorflow(module = NULL)
}
\arguments{
\item{module}{Name of sub-module to import. Defaults to \code{NULL}, which
imports the main tensorflow module.}
}
\value{
A tensorflow module
}
\description{
Import the tensorflow python module (or one of it's sub-modules) for
use in R.
}
\examples{
\dontrun{
tf <- tensorflow()
tflearn <- tensorflow("contrib.learn")
slim <- tenstensorflow("contrib.slim")
}

}

