% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{use_python}
\alias{use_python}
\alias{use_virtualenv}
\alias{use_condaenv}
\title{Configure which version of Python to use}
\usage{
use_python(python, required = FALSE)

use_virtualenv(virtualenv, required = FALSE)

use_condaenv(condaenv, conda = "auto", required = FALSE)
}
\arguments{
\item{python}{Path to Python binary}

\item{required}{Is this version of Python required? If \code{TRUE} then an error
occurs if it's not located. Otherwise, the version is taken as a hint only
and scanning for other versions will still proceed.}

\item{virtualenv}{Directory of Python virtualenv}

\item{condaenv}{Name of Conda environment}

\item{conda}{Conda executable. Default is \code{"auto"}, which checks the \code{PATH}
as well as other standard locations for Anaconda installations.}
}
\description{
Configure which version of Python to use
}
