% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualenv.R
\name{virtualenv-tools}
\alias{virtualenv-tools}
\alias{virtualenv_list}
\alias{virtualenv_create}
\alias{virtualenv_install}
\alias{virtualenv_remove}
\alias{virtualenv_root}
\alias{virtualenv_python}
\title{Interface to Python Virtual Environments}
\usage{
virtualenv_list()

virtualenv_create(envname, python = NULL)

virtualenv_install(envname, packages, ignored_installed = FALSE)

virtualenv_remove(envname, packages = NULL, confirm = interactive())

virtualenv_root()

virtualenv_python(envname)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment.}

\item{packages}{A character vector with package names to install or remove.}

\item{confirm}{Boolean; confirm before removing packages or virtual
environments?}

\item{ignore_installed}{Boolean; ignore previously-installed versions of the
requested packages?}
}
\description{
R functions for managing Python \href{https://virtualenv.pypa.io/en/stable/}{virtual environments}.
}
\details{
Virtual environments are by default located at \code{~/.virtualenvs}. You can change this
behavior by defining the \code{WORKON_HOME} environment variable.

Virtual environment functions are not supported on Windows (the use of
\link[=conda-tools]{conda environments} is recommended on Windows).
}
