% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{with-as-operator}
\alias{with-as-operator}
\alias{\%as\%}
\title{Create local alias for objects in \code{with} statements.}
\usage{
object \%as\% name
}
\arguments{
\item{object}{Object to alias}

\item{name}{Alias name}
}
\description{
Create local alias for objects in \code{with} statements.
}
\keyword{internal}
