% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_fn.R
\name{wrap_fn}
\alias{wrap_fn}
\title{Wrap an R function in a Python function with the same signature.}
\usage{
wrap_fn(f)
}
\arguments{
\item{f}{An R function}
}
\value{
A Python function that calls the R function \code{f} with the same signature.
}
\description{
This function could wrap an R function in a Python function with
the same signature. Note that this is function still experimental.
}
